[IAP/APPC-650] - implement a very basic payment with rewards with some mocked data

+++ appcoinsRewards/src/main/java/com/appcoins/wallet/appcoins/rewards/AppcoinsRewardsRepository.kt
@@ -4,11 +4,11 @@ import com.appcoins.wallet.appcoins.rewards.repository.bds.Origin
 import com.appcoins.wallet.appcoins.rewards.repository.bds.Type
 import io.reactivex.Completable
 import io.reactivex.Single
-import java.math.BigInteger
+import java.math.BigDecimal
 
 interface AppcoinsRewardsRepository {
   fun getBalance(address: String): Single<Long>
-  fun pay(walletAddress: String, signature: String, amount: BigInteger,
+  fun pay(walletAddress: String, signature: String, amount: BigDecimal,
           origin: Origin, sku: String,
           type: Type,
           developerAddress: String,
